The curl client seams not to work in the streaming client.
SimpleHTTPClient is again the default client. However, the
EventPublisher class explicitly uses the CurlHTTPClient.

Further investigation on the problem would help to now what the cause of
the problem is (probably, the curl library breaks the data it delivers
to the application at any point of the stream. If this is the case, the
application should be fixed (the parser should allow partial data).