Fixed an error detected when running the Twitter example.
The error message was: "error: cannot add/remove handle -
multi_perform() already running"

The cause seems to be that the asynchronous HTTP client cannot be used
to send requests from a callback run from that HTTP client. The
solution applied here was scheduling the fetch() operation of
EventPublisher to be executed after the callback from which the event
needs to be sent finishes. This way the final user has no need to know
about this restriction.