Configurable maximum number of clients in ztreamy.client.
Tornado's AsyncHTTPClient implementations may impose a
configurable maximum number of simultaneous fecth() operations
that can be performed on the same IOLoop. This number limits
the number of simultaneous streaming and long-polling clients.

The current default for this value in CurlAsyncHTTPClient is 10.
Applications that need clients for more than 10 simultaneous
streams must configure a bigger limit by calling the
new function configure_max_clients() in ztreamy.client.